refactor package backup
wal is in package backup
outsourcing variables and structs in seperate gofile
